Integrate conan into cmake
That includes moving the conanfile straight into cmake. CMake
should be able to configure and run conan automatically now.

We're also switching to the cmake_find_package conan
generator while we're at it. It seems to be the more 'modern'
way to do it now, and it does look cleaner and more flexible.