chore: add some Qt boilerplate
The build should now run MOC, RCC, and UIC automatically
as needed before compiling, and windeployqt at the end (on
Windows). macdeployqt will be added once I can test on a Mac.

win/macdeployqt is in QtTools, so it's added as a dependency.

Turns out that conan defaults to the static build of Qt, so
we're requesting a shared (.so/.dll) build explicitly. This also
allows win/macdeployqt to run properly.