refactor(Git): move Options to its own header (#20)
Git::Git is kind of becoming a kitchen sink, and that's bad for a number of reasons, including compilation speed. Git.hpp was shaping up to depend on literally everything in the library.

This PR also moves Git::Repository creation to static functions in the class, and does some more miscellaneous cleanup.