chore: upgrade Qt to 6.3 (#25)
6.4 is out, but it's not on conancenter yet.

I didn't change the CMakeLists file to use the new qt cmake functions because... ergh, it's going to take some more investigation. windeployqt only works in the install phase, which is a problem if I want to just run the app in the VS debugger... so that's going to take a separate ticket I think.