chore(CMakeLists): use the new CMake commands for Qt
These don't work as well as I'd like them to right now, but that's good
enough for a first commit, I guess.

Mainly, since the deploy scripts only run on install, I can't just start
the app in an IDE like Visual Studio and expect it to work, as it won't
find the Qt libraries. (This isn't a problem on Linux, but it is on
Windows and macOS.)